namespace fix, routing moved to MainBundle
