+++ app/AppKernel.php
@@ -16,7 +16,7 @@ class AppKernel extends Kernel
             new Symfony\Bundle\AsseticBundle\AsseticBundle(),
             new Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
             new Sensio\Bundle\FrameworkExtraBundle\SensioFrameworkExtraBundle(),
-            new Wizardry\WizardryMainBundle\WizardryMainBundle(),
+            new Wizardry\MainBundle\WizardryMainBundle(),
             new Doctrine\Bundle\MongoDBBundle\DoctrineMongoDBBundle(),
         );
