+++ app/config/routing.yml
@@ -1,3 +1,3 @@
-app:
-    resource: @MainBundle/Controller/
-    type:     annotation
+_wizardry:
+    resource: "@WizardryMainBundle/Resources/config/routing.yml"
+    prefix:   /
